refactor: move trace preview metadata types to homeboy-extension-contract
Clears one of the two TraceResults blockers. The TracePreviewMetadata type
closure (TracePreviewMetadata, TracePreviewAssetCheck,
TracePreviewAssetFanoutReport, TracePreviewAssetFanoutRequest,
TracePreviewLogPaths) is fully self-contained - 0 impls, no custom serde
helpers, external refs are only BTreeMap/Option/String.

The rig dependencies (TracePublicPreviewSpec etc.) that made preview.rs look
entangled live ONLY in the preview functions (start/start_external/
start_homeboy_native), not the metadata types. Data/behavior split: types move
to the contract crate, the preview-launch functions stay in core.

Core re-exports the types so all trace preview paths stay stable. Remaining
TraceResults blocker: RigStateSnapshot (crate::rig).
