+++ crates/homeboy-extension-contract/src/lib.rs
@@ -64,6 +64,7 @@ pub use manifest::ExtensionManifest;
 pub mod source_metadata_repair;
 pub mod test_drift;
 pub mod trace_config;
+pub mod trace_preview;
 pub mod update_output;
 pub mod version;
